requirements.txt
pyspark==3.2.4
mysql-connector-python==8.2.0
python-dotenv==1.0.1
pyyaml==6.0.1
pandas==1.5.3
